Move Person stream item code into concern
